Work on the "stop" button
Moved the "stop" button to its own SVG file, and set up the preview JS to let me switch the "stop" button state being previewed.

Added the icon label to the "stop" button. Added styles for the "disabled" state, and configured the glassy portions to only show on hover or when active. We still need an "active" state.