Hitbox fixes; "stop" button added to Interactive Preview
Our SVGs have rectangular hitboxes when used as images or background images, and `pointer-events` used within an SVG has no effect when the SVG is in static processing mode (i.e. when it's an image). As such, the lab's Interactive Preview has been updated: we use various tricks to force the buttons to have the correct hitbox, including by moving the visuals to pseudo-elements.

Added the "stop" button to the Interactive Preview.

Removed empty space from the "stop" button's bounds.