Experimenting with easier ways to define "fuzzy" shapes
Fuzzy-rect update: simplified the corner and edge styles by rotating the entire corner or edge shape about the rect center. This means we need fewer edge gradient definitions.

Fuzzy-rect update: updated the gradients to go from white to black, for use in a <mask/>, instead of going from opaque to transparent. The CSS `color-interpolation` property only affects RGB, not A, so opaque-to-transparent gradients don't benefit from it but white-to-black gradients do. The CSS calculations for converting from sRGB to linear color stops are still present, but I will likely remove them soon.