"Next" button icon
This button can switch between "next" and "fast-forward" icons, so we'll need to implement that. The current implementation should offer a foundation for that.

However, I've noticed that artifacts can appear when the buttons change state. It's as if the browser actually goes out and re-fetches the SVG, even though only the URI fragment is changing (and even though the SVG file is local!). I may want to change my approach: instead of having the distinct elements of the SVG file actually change, it may be wiser to instead use a spritesheet approach.