"Next" button work
Converted the "next" button into a spritesheet; work will proceed there from now on. This allows us to significantly reduce the amount of CSS we're using within the SVG, which is good for compatibility with non-browsers.

Set up a more precise hitbox for the "next" button in the lab's Interactive Preview.

Finished diffuse for the "next" button's "active" state; we just need alpha and this button'll be done!