More work on the WMPlayerElement
Fixed the issue with the slider thumbs on Chrome; it was a result of the size attributes on the SVG.

The scaling factor for the entire player is now adjustable via a `--scale` CSS variable.

The "stop" button is now initially disabled, and becomes disabled when used (until you play the media again).