Shuffle, loop, stop glyphs; WMPlayerElement work
Created glyphs for the Shuffle, Loop, and Stop buttons, and put them in a spritesheet.

Split the WMPlayerSliderElement into its own JS file.

Started work on adding playlist functionality to the WMPlayerElement, and on cleaning up the code generally.

Added tooltips to the Loop and Shuffle buttons.

Updated the WMPlayerElement CSS to afford greater control over the video sizing, and to letterbox/windowbox the video when it's smaller than the available display area.