Various bugfixes
Focus borders for the skinned buttons. The tricks we use to size buttons properly without giving them box-shaped hitboxes result in the native focus borders being lost. These replacements should hopefully be good enough.

ARIA attributes and keyboard controls for the custom sliders.

Fixed a bug wherein a custom element could be created, have properties set on it, and only after that be upgraded, such that attempts to use a setter could instead create an instance-level expando with the same name. I witnessed this happening on the sliders, even despite the slider script running (and defining the custom element) before any code which would create a slider, but I've added the fix to both custom elements just in case.

Fixed a typo wherein changing the volume slider would set the volume based on the current value of the seek slider.

The volume slider now stores values as [0, 100] rather than [0, 1]. If we can figure out how to implement showing the value as a tooltip similarly to how WMP does it (constantly visible tooltip shown during editing), this will be helpful. It's also a better way to define the step values for keyboard control.

The custom sliders are now form-associated. We don't need this, but it seems fair to set it up since the sliders are basically a replacement for input[type=range].

The custom sliders can now be disabled, though we have no test or UI for this yet.