Previous can now be Rewind-only; bugfixes
If we can't go back to the previous media item (e.g. because there isn't one), but we can rewind, then the Previous button now shows the Rewind glyph only, and is not disabled.

Changed how we initiate rewinding so that we don't get an extra delay on holding the button down.

We now update the Previous button when the seek slider changes, and not just in response to current-time-change events. This ensures that we update the button immediately and not after a brief delay (e.g. if the media needs to buffer).

Fixed mistakes with dispatching CustomEvents from the playlist class. These were causing a new bug wherein the Next button would fail to update its enable state when you toggled looping or shuffling on and off.