Critical shuffle fix; code cleanup; minor bugfix
Critical fix to playlists: when shuffling, if the current item is the only remaining shuffle item, then skip it rather than getting trapped in the retry loop.

Certain interactions with the player will disqualify it from autoplaying, so that if it takes too long to start autoplaying, it doesn't interrupt anything the user has already started doing. The disqualification function is no longer invoked individually by each interaction handler (e.g. the click handler for the play/pause button, etc.), but rather instead has been made able to function as an event listener, and is attached as a separate event handler to each such widget.

Cleaned up handling of the Stop button and its state.

The Previous and Next buttons were originally coded to react to the click event but later changed to react to mousedown and mouseup. The click event was amended, long ago, to only fire for the primary mouse button, but the individual mouse events fire for all buttons; the Previous and Next handlers now check the mouse button that triggered their events.