Additional graphical work
When in theater mode (designed to match WMP's Now Playing view), we now use white button glyphs, consistent with WMP. This fix is bound to using CSS background-position rather than SVG views for the glyphs (see below), as for simplicity and to keep filesizes small, we co-opt the fragment identifier and use in-SVG stylesheets to toggle the icon colors.

FIX: If the player is initially paused and you click Next, we now properly update the enable state of the Previous button.

Updated the buttons and their glyphs to use CSS background-position rather than SVG views. This means that browsers will no longer send additional network requests for each state change (which they apparently do, as can be verified in Chrome's network debugger, even if the URL change just involves the fragment identifier).

Renamed the glyph spritesheet for the Play/Pause button, for consistency with the other glyph spritesheets' filenames.