Dark theme improvements
The player now uses an internal wrapper `div` for its layout, so setting properties like `display` on the player itself (i.e. from the light DOM) won't blow away half the layout. This has the additional benefit of (more reliably) allowing us to attach pseudo-elements for the dark theme background to that wrapper, such that they can be sized and positioned based entirely on the grid.

The dark theme now uses the same tray spritesheet as the overlaid controls, but with a different color palette. (That said, I think the tray metrics differ between Now Playing and Library, separately from the former's 92% scaling factor, so I'll likely need a different spritesheet. I can design it the same way, though.)