Tweaked button glass for the Prev button (sub-pixels)
The Previous button is basically just the Next button but rotated 180 degrees about its center. When a button's glass sprite has a non-integer position, we use a 3D transform and some other trickery to try and improve how browsers perform sub-pixel rounding. These two facts in combination were causing sub-pixel rounding errors for this specific button in Firefox, when controls are overlaid (such that the sprite being rounded was shown at 92% scale), vertically displacing the sprite.

The sub-pixel transform was applied to a pseudo-element, but the rotation was applied to the entire button. The solution was to make it so that the Previous button, specifically, clears the transform off the pseudo-element and applies said transform itself in tandem with the rotation.