We now allow rearranging player controls via an API
Fixed the Play/Pause button using the wrong CSS variable to specify its current glyph.

The player now exposes attributes (with reflecting JS properties) for configuring which controls are visible in the tray. You can use this to choose which buttons are visible, and to move the seek slider into the controls tray next to the Play/Pause button.