Critical fixes for data forwarding; testing text tracks
Fixed forwarding of various getters, setters, and attributes from the custom element to the wrapped media element. We were accidentally adding static getters and setters to the constructor, rather than to the prototype.

Fixed the playlist's handling of HTMLTrackElements.

Added a test case for an audio file with a lyrics track. (Test files not included, as they are copyrighted music and the lyrics thereof.)