Player fixes and demo site
FIX: The player JS now tries to figure out its own filepath, so that if it's in a subfolder relative to the current page, the CSS is still included properly.

Started work on the demo site.

FIX: Moving the seek slider into the righthand gutter no longer compresses it to zero-width.

FIX: The Previous and Next buttons now theme properly when data-overlay-controls="video-only".