Various fixes
FIX: Guarding against NaN slider operations seen during testing. Utterly mystified as to what could've caused them.

FIX: Sliders should no longer react to swipe touch-inputs used to scroll the page.

FIX: Tweaked layout for the current timestamp and some of the spacing between specific button types. The goal for the former is to improve the timestamp's position at narrow display widths; the goal of the latter is to get the left and right "wings" of the tray to be closer to equal size when showing controls in their default configuration.

There remain issues with narrow display widths. In essence, the controls, and the contents of the left and right gutters, can only shrink so much, horizontally, and so if the page isn't wide enough for them, then they will overflow, unavoidably. (We can set the meta viewport width to a known safe value, but then the player buttons end up... rather small.) I'm exploring what options I have to remedy that, e.g. by using container queries to move the gutters upward, but it's not looking good. (How, for example, would I specifically know that the controls tray is too wide for the gutters to fit alongside it?)